Implemented basic Login
At the Login screen, the user must fill in both fields in order to press "Login". If they have filled these in incorrectly, nothing happens (other than a Debug.Log). If they are filled in correctly, the Client's Player object is assigned the correct tokens and it moves on to the Lobby Screen